Refactor
Refactored awards for v1, v2. Added views folder to v2.
